+++ sdk/src/main/java/com/ciscowebex/androidsdk/phone/internal/CallMembershipImpl.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2020 Cisco Systems Inc
+ * Copyright 2016-2021 Cisco Systems Inc
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -22,15 +22,20 @@
 
 package com.ciscowebex.androidsdk.phone.internal;
 
-import com.ciscowebex.androidsdk.internal.model.*;
+import com.ciscowebex.androidsdk.internal.model.FloorModel;
+import com.ciscowebex.androidsdk.internal.model.LocusMediaDirection;
+import com.ciscowebex.androidsdk.internal.model.LocusModel;
+import com.ciscowebex.androidsdk.internal.model.LocusParticipantModel;
 import com.ciscowebex.androidsdk.phone.CallMembership;
 import com.ciscowebex.androidsdk.utils.WebexId;
 import com.github.benoitdion.ln.Ln;
 
-import me.helloworld.utils.Checker;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
+
+import me.helloworld.utils.Checker;
 
 public class CallMembershipImpl implements CallMembership {
 
@@ -65,7 +70,7 @@ public class CallMembershipImpl implements CallMembership {
     CallMembershipImpl(LocusParticipantModel participant, CallImpl call) {
         this.call = call;
         setModel(participant);
-        Ln.d("CallMembership: " + getId() + " person: " + getPersonId() + " email: " + getEmail() + "  video: " +  isSendingVideo() + "   audio: " + isSendingAudio());
+        Ln.d("CallMembership: " + getId() + " person: " + getPersonId() + " email: " + getEmail() + "  video: " + isSendingVideo() + "   audio: " + isSendingAudio());
     }
 
     public String getId() {
@@ -120,6 +125,24 @@ public class CallMembershipImpl implements CallMembership {
         return Checker.isEqual(getId(), ((CallMembershipImpl) call.getActiveSpeaker()).getId());
     }
 
+    @Override
+    public boolean isAudioMutedControlled() {
+        return model.isAudioRemotelyMuted();
+    }
+
+    @Override
+    @Nullable
+    public String audioModifiedBy() {
+        String personId = null;
+        if (model.getControls() != null && model.getControls().getAudio() != null && model.getControls().getAudio().getMeta() != null) {
+            personId = model.getControls().getAudio().getMeta().getModifiedBy();
+        }
+        if (personId != null) {
+            personId = new WebexId(WebexId.Type.PEOPLE, WebexId.DEFAULT_CLUSTER, personId).getBase64Id();
+        }
+        return personId;
+    }
+
     public LocusParticipantModel getModel() {
         synchronized (this) {
             return model;
@@ -149,7 +172,7 @@ public class CallMembershipImpl implements CallMembership {
                 + " isSelf: " + isSelf()
                 + " person: " + getPersonId()
                 + " email: " + getEmail()
-                + " video: " +  isSendingVideo()
+                + " video: " + isSendingVideo()
                 + " audio: " + isSendingAudio();
     }