+++ sdk/src/main/java/com/ciscowebex/androidsdk/phone/internal/PhoneImpl.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2020 Cisco Systems Inc
+ * Copyright 2016-2021 Cisco Systems Inc
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -33,30 +33,60 @@ import android.util.Pair;
 import android.view.View;
 
 import com.cisco.wme.appshare.ScreenShareContext;
-import com.ciscowebex.androidsdk.*;
+import com.ciscowebex.androidsdk.CompletionHandler;
+import com.ciscowebex.androidsdk.Webex;
+import com.ciscowebex.androidsdk.WebexError;
+import com.ciscowebex.androidsdk.WebexEvent;
 import com.ciscowebex.androidsdk.auth.Authenticator;
-import com.ciscowebex.androidsdk.internal.*;
+import com.ciscowebex.androidsdk.internal.AcquirePermissionActivity;
+import com.ciscowebex.androidsdk.internal.Credentials;
+import com.ciscowebex.androidsdk.internal.Device;
+import com.ciscowebex.androidsdk.internal.MetricsService;
+import com.ciscowebex.androidsdk.internal.ResultImpl;
 import com.ciscowebex.androidsdk.internal.crypto.KeyManager;
-import com.ciscowebex.androidsdk.internal.mercury.*;
-import com.ciscowebex.androidsdk.internal.metric.CallAnalyzerReporter;
-import com.ciscowebex.androidsdk.internal.queue.Queue;
 import com.ciscowebex.androidsdk.internal.media.MediaCapability;
 import com.ciscowebex.androidsdk.internal.media.WMEngine;
-import com.ciscowebex.androidsdk.internal.model.*;
+import com.ciscowebex.androidsdk.internal.mercury.MercuryActivityEvent;
+import com.ciscowebex.androidsdk.internal.mercury.MercuryEvent;
+import com.ciscowebex.androidsdk.internal.mercury.MercuryKmsMessageEvent;
+import com.ciscowebex.androidsdk.internal.mercury.MercuryLocusEvent;
+import com.ciscowebex.androidsdk.internal.mercury.MercuryService;
+import com.ciscowebex.androidsdk.internal.metric.CallAnalyzerReporter;
+import com.ciscowebex.androidsdk.internal.model.ActivityModel;
+import com.ciscowebex.androidsdk.internal.model.FloorModel;
+import com.ciscowebex.androidsdk.internal.model.KmsMessageModel;
+import com.ciscowebex.androidsdk.internal.model.LocusModel;
+import com.ciscowebex.androidsdk.internal.model.LocusSequenceModel;
+import com.ciscowebex.androidsdk.internal.model.MediaEngineReachabilityModel;
+import com.ciscowebex.androidsdk.internal.model.MediaInfoModel;
+import com.ciscowebex.androidsdk.internal.model.MediaShareModel;
+import com.ciscowebex.androidsdk.internal.model.ObjectModel;
+import com.ciscowebex.androidsdk.internal.queue.Queue;
 import com.ciscowebex.androidsdk.internal.reachability.BackgroundChecker;
 import com.ciscowebex.androidsdk.membership.MembershipObserver;
 import com.ciscowebex.androidsdk.membership.internal.InternalMembership;
 import com.ciscowebex.androidsdk.message.MessageObserver;
 import com.ciscowebex.androidsdk.message.internal.InternalMessage;
 import com.ciscowebex.androidsdk.message.internal.MessageClientImpl;
-import com.ciscowebex.androidsdk.phone.*;
+import com.ciscowebex.androidsdk.phone.AdvancedSetting;
+import com.ciscowebex.androidsdk.phone.Call;
+import com.ciscowebex.androidsdk.phone.CallMembership;
+import com.ciscowebex.androidsdk.phone.CallObserver;
+import com.ciscowebex.androidsdk.phone.MediaOption;
+import com.ciscowebex.androidsdk.phone.Phone;
 import com.ciscowebex.androidsdk.space.SpaceObserver;
 import com.ciscowebex.androidsdk.space.internal.InternalSpace;
 import com.ciscowebex.androidsdk.utils.Utils;
 import com.ciscowebex.androidsdk.utils.WebexId;
 import com.github.benoitdion.ln.Ln;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.UUID;
 
 public class PhoneImpl implements Phone, UIEventHandler.EventObserver, MercuryService.MercuryListener, BackgroundChecker.BackgroundListener {
 
@@ -76,6 +106,7 @@ public class PhoneImpl implements Phone, UIEventHandler.EventObserver, MercurySe
     private State state = State.UNREGISTERED;
     private MediaEngine engine;
     private Map<String, CallImpl> calls = new HashMap<>();
+    private List<String> activeSpaceIds = new ArrayList<>();
     private CallContext callContext;
     private final CallService service;
     private final ReachabilityService reachability;
@@ -92,6 +123,8 @@ public class PhoneImpl implements Phone, UIEventHandler.EventObserver, MercurySe
     private List<String> audioEnhancementModels = null;
     private Map<Class<? extends AdvancedSetting>, AdvancedSetting> settings = new HashMap<>();
     private boolean enableBackgroundStream = false;
+    private boolean enableAudioBNR = false;
+    private AudioBRNMode audioBRNMode = AudioBRNMode.HP;
 
     private String uuid = UUID.randomUUID().toString();
     private boolean canceled = false;
@@ -325,12 +358,10 @@ public class PhoneImpl implements Phone, UIEventHandler.EventObserver, MercurySe
                     Queue.main.run(this::tryAcquirePermission);
                     Queue.serial.yield();
                 });
-            }
-            else if (action == H264LicenseAction.DECLINE) {
+            } else if (action == H264LicenseAction.DECLINE) {
                 Ln.d("Decline H264 license");
                 ResultImpl.errorInMain(callback, WebexError.from(WebexError.ErrorCode.DECLINE_H264_LICENSE));
-            }
-            else if (action == H264LicenseAction.VIEW_LICENSE) {
+            } else if (action == H264LicenseAction.VIEW_LICENSE) {
                 Ln.d("View H264 license");
                 Intent browserIntent = new Intent(Intent.ACTION_VIEW, Uri.parse(H264LicensePrompter.LICENSE_URL));
                 browserIntent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
@@ -386,7 +417,7 @@ public class PhoneImpl implements Phone, UIEventHandler.EventObserver, MercurySe
                     String correlationId = UUID.randomUUID().toString();
                     //CallAnalyzerReporter.shared.reportJoinRequest(correlationId, null);
                     if (target instanceof CallService.CallableTarget) {
-                        service.call(((CallService.CallableTarget) target).getCallee(), outgoing.getOption(), correlationId, device, localSdp,reachabilities, callResult -> {
+                        service.call(((CallService.CallableTarget) target).getCallee(), outgoing.getOption(), correlationId, device, localSdp, reachabilities, callResult -> {
                             if (callResult.getError() != null || callResult.getData() == null) {
                                 Queue.main.run(() -> outgoing.getCallback().onComplete(ResultImpl.error(callResult.getError())));
                                 Queue.serial.yield();
@@ -395,7 +426,7 @@ public class PhoneImpl implements Phone, UIEventHandler.EventObserver, MercurySe
                             }
                             doLocusResponse(new LocusResponse.Call(device, correlationId, session, callResult.getData(), outgoing.getCallback()), Queue.serial);
                         });
-                    } else if (target instanceof CallService.JoinableTarget){
+                    } else if (target instanceof CallService.JoinableTarget) {
                         service.getOrCreatePermanentLocus(((CallService.JoinableTarget) target).getConversation(), device, convResult -> {
                             if (this.canceled) {
                                 this.canceled = false;
@@ -419,8 +450,7 @@ public class PhoneImpl implements Phone, UIEventHandler.EventObserver, MercurySe
                                 doLocusResponse(new LocusResponse.Call(device, correlationId, session, joinResult.getData(), outgoing.getCallback()), Queue.serial);
                             });
                         });
-                    }
-                    else {
+                    } else {
                         Ln.e("Cannot find dial target: " + outgoing.getTarget());
                         Queue.main.run(() -> outgoing.getCallback().onComplete(ResultImpl.error("Cannot find dial target: " + outgoing.getTarget())));
                         Queue.serial.yield();
@@ -537,6 +567,26 @@ public class PhoneImpl implements Phone, UIEventHandler.EventObserver, MercurySe
     }
 
     @Override
+    public void enableAudioBNR(boolean enable) {
+        this.enableAudioBNR = enable;
+    }
+
+    @Override
+    public boolean isAudioBNREnable() {
+        return enableAudioBNR;
+    }
+
+    @Override
+    public void setAudioBNRMode(AudioBRNMode mode) {
+        this.audioBRNMode = mode;
+    }
+
+    @Override
+    public AudioBRNMode getAudioBNRMode() {
+        return audioBRNMode;
+    }
+
+    @Override
     public void requestVideoCodecActivation(@NonNull AlertDialog.Builder builder, @Nullable CompletionHandler<H264LicenseAction> callback) {
         this.prompter.check(getContext(), builder, result -> {
             if (callback == null) {
@@ -544,8 +594,7 @@ public class PhoneImpl implements Phone, UIEventHandler.EventObserver, MercurySe
                     Intent browserIntent = new Intent(Intent.ACTION_VIEW, Uri.parse(H264LicensePrompter.LICENSE_URL));
                     this.getContext().startActivity(browserIntent);
                 }
-            }
-            else {
+            } else {
                 callback.onComplete(result);
             }
         });
@@ -753,14 +802,13 @@ public class PhoneImpl implements Phone, UIEventHandler.EventObserver, MercurySe
                     Queue.main.run(this::tryAcquirePermission);
                     Queue.serial.yield();
                 });
-            }
-            else if (action == H264LicenseAction.DECLINE) {
+            } else if (action == H264LicenseAction.DECLINE) {
                 Ln.d("Decline H264 license");
                 ResultImpl.errorInMain(callback, WebexError.from(WebexError.ErrorCode.DECLINE_H264_LICENSE));
-            }
-            else if (action == H264LicenseAction.VIEW_LICENSE) {
+            } else if (action == H264LicenseAction.VIEW_LICENSE) {
                 Ln.d("View H264 license");
                 Intent browserIntent = new Intent(Intent.ACTION_VIEW, Uri.parse(H264LicensePrompter.LICENSE_URL));
+                browserIntent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
                 this.getContext().startActivity(browserIntent);
                 ResultImpl.errorInMain(callback, WebexError.from(WebexError.ErrorCode.VIEW_H264_LICENSE));
             }
@@ -833,8 +881,7 @@ public class PhoneImpl implements Phone, UIEventHandler.EventObserver, MercurySe
                     if (error != null) {
                         if (error.is(WebexError.ErrorCode.CONFLICT_ERROR) || error.is(WebexError.ErrorCode.NETWORK_ERROR)) {
                             call.end(new CallObserver.CallErrorEvent(call, error));
-                        }
-                        else {
+                        } else {
                             Queue.main.run(() -> callback.onComplete(ResultImpl.error(error)));
                             Queue.serial.yield();
                             return;
@@ -853,8 +900,7 @@ public class PhoneImpl implements Phone, UIEventHandler.EventObserver, MercurySe
             if (url == null) {
                 Ln.e("Missing self participant URL");
                 Queue.serial.yield();
-            }
-            else {
+            } else {
                 service.layout(url, device, layout, result -> Queue.serial.yield());
             }
         });
@@ -1128,6 +1174,17 @@ public class PhoneImpl implements Phone, UIEventHandler.EventObserver, MercurySe
         Ln.d("doLocusEvent: " + url);
         //CallAnalyzerReporter.shared.reportClientNotificationReceived(model.getKey(), true);
         Queue.serial.run(() -> {
+            if (!model.isOneOnOne() && model.isValid() && model.getConversationUrl() != null) {
+                String spaceId = WebexId.from(model.getConversationUrl(), device).getBase64Id();
+                if (!activeSpaceIds.contains(spaceId) && model.getFullState() != null && model.getFullState().isActive()) {
+                    fire(new InternalSpace.InternalSpaceCallStarted(spaceId, model));
+                    activeSpaceIds.add(spaceId);
+                }
+                if (activeSpaceIds.contains(spaceId) && model.getFullState() != null && !model.getFullState().isActive()) {
+                    fire(new InternalSpace.InternalSpaceCallEnded(spaceId, model));
+                    activeSpaceIds.remove(spaceId);
+                }
+            }
             CallImpl call = calls.get(url);
             if (call == null) {
                 if (device != null && model.isIncomingCall() && model.isValid()) {
@@ -1176,51 +1233,42 @@ public class PhoneImpl implements Phone, UIEventHandler.EventObserver, MercurySe
                 && object != null && object.isPerson()
                 && target != null && target.isConversation()) {
             fire(new InternalMembership.InternalMembershipCreated(new InternalMembership(activity, clusterId), activity));
-        }
-        else if (verb == ActivityModel.Verb.leave
+        } else if (verb == ActivityModel.Verb.leave
                 && (object == null || object.isPerson())
                 && target != null && target.isConversation()) {
             fire(new InternalMembership.InternalMembershipDeleted(new InternalMembership(activity, clusterId), activity));
-        }
-        else if ((activity.getVerb() == ActivityModel.Verb.assignModerator || activity.getVerb() == ActivityModel.Verb.unassignModerator)
+        } else if ((activity.getVerb() == ActivityModel.Verb.assignModerator || activity.getVerb() == ActivityModel.Verb.unassignModerator)
                 && object != null && object.isPerson()
                 && target != null && target.isConversation()) {
             fire(new InternalMembership.InternalMembershipUpdated(new InternalMembership(activity, clusterId), activity));
-        }
-        else if (activity.getVerb() == ActivityModel.Verb.acknowledge
+        } else if (activity.getVerb() == ActivityModel.Verb.acknowledge
                 && object != null && object.isActivity() && object.getId() != null
                 && target != null && target.isConversation()) {
             fire(new InternalMembership.InternalMembershipMessageSeen(new InternalMembership(activity, clusterId), activity, new WebexId(WebexId.Type.MESSAGE, clusterId, object.getId()).getBase64Id()));
-        }
-        else if (activity.getVerb() == ActivityModel.Verb.create
+        } else if (activity.getVerb() == ActivityModel.Verb.create
                 && object != null && object.isConversation() && object.getId() != null) {
             String base64Id = new WebexId(WebexId.Type.ROOM, clusterId, object.getId()).getBase64Id();
-            webex.spaces().get(base64Id, result -> fire(result.getData() == null ? null : new InternalSpace.InternalSpaceCeated(result.getData(), activity)));
-        }
-        else if (activity.getVerb() == ActivityModel.Verb.update
+            webex.spaces().get(base64Id, result -> fire(result.getData() == null ? null : new InternalSpace.InternalSpaceCreated(result.getData(), activity)));
+        } else if (activity.getVerb() == ActivityModel.Verb.update
                 && object != null && object.isConversation()
                 && target != null && target.isConversation() && target.getId() != null) {
             String base64Id = new WebexId(WebexId.Type.ROOM, clusterId, object.getId()).getBase64Id();
             webex.spaces().get(base64Id, result -> fire(result.getData() == null ? null : new InternalSpace.InternalSpaceUpdated(result.getData(), activity)));
-        }
-        else if ((activity.getVerb() == ActivityModel.Verb.post || activity.getVerb() == ActivityModel.Verb.share)
+        } else if ((activity.getVerb() == ActivityModel.Verb.post || activity.getVerb() == ActivityModel.Verb.share)
                 && activity.getConversationId() != null && activity.getConversationUrl() != null) {
             ((MessageClientImpl) webex.messages()).doMessageReveived(activity, clusterId, message -> {
                 fire(new InternalMessage.InternalMessageReceived(message, activity));
                 // TODO Remove the deprecated event in next big release
                 fire(new InternalMessage.InternalMessageArrived(message, activity));
             });
-         }
-        else if (activity.getVerb() == ActivityModel.Verb.update
+        } else if (activity.getVerb() == ActivityModel.Verb.update
                 && activity.getConversationId() != null && activity.getConversationUrl() != null
                 && object != null && object.isContent() && object.getId() != null) {
             ((MessageClientImpl) webex.messages()).doMessageUpdated(activity, this::fire);
-        }
-        else if (activity.getVerb() == ActivityModel.Verb.delete
+        } else if (activity.getVerb() == ActivityModel.Verb.delete
                 && object != null && object.isActivity() && object.getId() != null) {
             fire(new InternalMessage.InternalMessageDeleted(((MessageClientImpl) webex.messages()).doMessageDeleted(object.getId(), clusterId), activity));
-        }
-        else {
+        } else {
             Ln.d("Not a valid activity: " + url);
         }
     }
@@ -1229,11 +1277,9 @@ public class PhoneImpl implements Phone, UIEventHandler.EventObserver, MercurySe
         Queue.main.run(() -> {
             if (event instanceof MembershipObserver.MembershipEvent && membershipObserver != null) {
                 membershipObserver.onEvent((MembershipObserver.MembershipEvent) event);
-            }
-            else if (event instanceof SpaceObserver.SpaceEvent && spaceObserver != null) {
+            } else if (event instanceof SpaceObserver.SpaceEvent && spaceObserver != null) {
                 spaceObserver.onEvent((SpaceObserver.SpaceEvent) event);
-            }
-            else if (event instanceof MessageObserver.MessageEvent && messageObserver != null) {
+            } else if (event instanceof MessageObserver.MessageEvent && messageObserver != null) {
                 messageObserver.onEvent((MessageObserver.MessageEvent) event);
             }
         });
@@ -1277,6 +1323,8 @@ public class PhoneImpl implements Phone, UIEventHandler.EventObserver, MercurySe
         capability.setHardwareCodecEnable(isHardwareAccelerationEnabled());
         capability.setHardwareVideoSetting(getHardwareVideoSettings());
         capability.setAudioEnhancementModels(audioEnhancementModels);
+        capability.setEnableAudioBNR(isAudioBNREnable());
+        capability.setAudioBRNMode(getAudioBNRMode());
         if (device != null) {
             capability.setDeviceSettings(device.getDeviceSettings());
         }